requirements/base.txt

# Database Adapter
MySQL-python


# Django stuff
Django

# Templates
django_compressor

# Celery: Message queue
celery
django-celery

# L10n/i18n
#Babel
#-e git://github.com/clouserw/tower.git#egg=tower

# Caching
python-memcached

# Logging
logutils

# Mixins
django-braces
django-model-utils
bpython

# Admin
django-debug-toolbar

# utils
path.py
dj-database-url
django-extensions
django-widget-tweaks

# rest api
djangorestframework
markdown
django-filter

# angular
django-angular
